requirements.txt
asgiref==3.5.0
Django==4.0.3
django-extensions==3.1.5
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.1
